Convert to size_t and formatting

+++ src/groups/mwc/mwctst/mwctst_testhelper.cpp
@@ -47,7 +47,7 @@ void TestHelper::printTestName(bslstl::StringRef value)
     }
 
     bsl::cout << "\n" << value << "\n";
-    long unsigned int length = value.length();
+    size_t length = value.length();
     while (length--) {
         bsl::cout << "=";
     }
